backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
python-dotenv==1.1.0
httpx==0.28.1
pyjwt==2.10.1
apscheduler==3.11.0
python-multipart==0.0.20
